% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSData.R
\docType{data}
\name{KSData}
\alias{KSData}
\title{Kinase-Substrate (K-S) Relationship Dataset}
\format{abbreviated dataframe containing the kinase-substrate annotations and source}
\usage{
data(KSData)
}
\description{
K-S annotations from PhosphoSitePlus and NetworKIN predictions;
This is an abbreviated version of the full dataset used purely for demonstration;
please go to the GitHub page for access to the complete file: github.com/casecpb/KSEA/
}
\references{
Hornbeck et al. (2015) Nucleic Acids Res. 43:D512-20

Horn et al. (2014) Nature Methods 11(6):603-4
}
\keyword{datasets}
